refactor(service): Update payment method ID type from int to long
- Updated `IsPaymentMethodNameExistsAsync()` in `IPaymentMethodService`:
  - Changed parameter type from `int id` to `long id`.
- Updated method implementation in `PaymentMethodService` accordingly.
- Ensures compatibility with long ID values for better scalability.

This refactor improves the service layer by allowing support for larger ID values.

+++ backend/Service/Interfaces/IPaymentMethodService.cs
@@ -7,5 +7,5 @@ public interface IPaymentMethodService : IGenericService<PaymentMethod>
 {
     public Task<PaymentMethod?> GetPaymentMethodByNameAsync(string name);
     public Task<bool> IsPaymentMethodNameExistsAsync(string name);
-    public Task<bool> IsPaymentMethodNameExistsAsync(string name, int id);
+    public Task<bool> IsPaymentMethodNameExistsAsync(string name, long id);
 }
